Removed the need to redraw all elements every move
using more intelligent CSS and correct formatting removes the need to
redraw the page, instead just move the cards from one place to another.
This nicely resulted also in the drag image accidentally also showing
cards on top of the card being dragged when you drag a stack of cards of
same suit.