Implemented option for 1 or 2 suits
You can select the level (number of suits) for the next game in the menu
dialog, if you are not in the middle of a game already then it will
re-deal the cards for you using your new "level", otherwise it will just
remember it for next time you start a new game (but will forget that if
you don't finish the game before closing the browser).  The high scores
take the level in to account: higher levels are higher up the board
regardless of the score.  This addresses issue #1